修复因网络验证返回结构的变更导致验证失败的问题；修复血点模式不移动的问题

+++ Utils/GameOperate.py
@@ -123,7 +123,7 @@ def killer_ctrl(skill_key: str = 'lcontrol', skill_time: float = 4.3) -> None:
     key_up(skill_key)
     time.sleep(0.1)
     key_down(skill_key)
-    for _ in range(skill_time*10):
+    for _ in range(int(skill_time*10)):
         time.sleep(0.1)
     key_up(skill_key)
